+++ .husky/pre-commit
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-npm run lint:fix && npm run format && npm run typecheck
+npm run db:generate-type && npm run lint:fix && npm run format && npm run typecheck
