+++ app/modules/asset/types.ts
@@ -1,4 +1,9 @@
-import type { AssetCustomFieldValue } from "@prisma/client";
+import type {
+  Asset,
+  AssetCustomFieldValue,
+  CustomField,
+  User,
+} from "@prisma/client";
 
 export interface ICustomFieldValueJson {
   raw: string | number | boolean;
@@ -13,3 +18,62 @@ export type ShelfAssetCustomFieldValueType = Omit<
   AssetCustomFieldValue,
   "value"
 > & { value: ICustomFieldValueJson };
+
+export interface UpdateAssetPayload {
+  id: Asset["id"];
+  title?: Asset["title"];
+  description?: Asset["description"];
+  categoryId?: Asset["categoryId"];
+  newLocationId?: Asset["locationId"];
+  currentLocationId?: Asset["locationId"];
+  mainImage?: Asset["mainImage"];
+  mainImageExpiration?: Asset["mainImageExpiration"];
+  tags?: { set: { id: string }[] };
+  userId?: User["id"];
+  customFieldsValues?: ShelfAssetCustomFieldValueType[];
+}
+
+export interface CreateAssetFromContentImportPayload
+  extends Record<string, any> {
+  title: string;
+  description?: string;
+  category?: string;
+  tags: string[];
+  location?: string;
+  custodian?: string;
+}
+export interface CreateAssetFromBackupImportPayload
+  extends Record<string, any> {
+  id: string;
+  title: string;
+  description?: string;
+  category:
+    | {
+        id: string;
+        name: string;
+        description: string;
+        color: string;
+        createdAt: string;
+        updatedAt: string;
+        userId: string;
+      }
+    | {};
+  tags: {
+    name: string;
+  }[];
+  location:
+    | {
+        name: string;
+        description?: string;
+        address?: string;
+        createdAt: string;
+        updatedAt: string;
+      }
+    | {};
+  customFields: AssetCustomFieldsValuesWithFields[];
+}
+
+export type AssetCustomFieldsValuesWithFields =
+  ShelfAssetCustomFieldValueType & {
+    customField: CustomField;
+  };
